autopkgtest: Run autopkgtest-virt-schroot as user, not root
Not all packages' tests cope well with having uid 0. For example,
GLib has tests that assert that it sees EACCES on trying to chdir
into a non-traversable directory, an assumption that fails if it
has CAP_DAC_OVERRIDE.

To support this, make transient directories world-readable, make
the autopkgtest output directory world-writeable, and upload
source and binary files to the schroot test worker owned by
the target user (so it can hard-link them into place).